refactor: move gcs logic from shared to mirror-schema
