+++ mirror/mirror-schema/src/cloud-storage.test.ts
@@ -1,12 +1,20 @@
 import {expect, test} from '@jest/globals';
-import {parseCloudStorageURL} from './parse-cloud-storage-url.js';
+import {parseCloudStorageURL} from './cloud-storage.js';
 
 test('parseCloudStorageURL', () => {
   expect(parseCloudStorageURL('gs://bucket-name/filename')).toEqual({
     bucketName: 'bucket-name',
     filename: 'filename',
   });
+  expect(parseCloudStorageURL('gcs://bucket-name/filename')).toEqual({
+    bucketName: 'bucket-name',
+    filename: 'filename',
+  });
   expect(parseCloudStorageURL('gs://bucket-name/path/filename')).toEqual({
+    bucketName: 'bucket-name',
+    filename: 'path/filename',
+  });
+  expect(parseCloudStorageURL('gcs://bucket-name/path/filename')).toEqual({
     bucketName: 'bucket-name',
     filename: 'path/filename',
   });
