+++ mirror/mirror-schema/src/cloud-storage.ts
@@ -6,7 +6,7 @@ export function parseCloudStorageURL(url: string): {
 } {
   // Don't use the built in URL class, because Node.js implementation is not
   // standards compliant.
-  const m = url.match(/^gs:\/\/([^/]+)\/(.+)$/);
+  const m = url.match(/^gc?s:\/\/([^/]+)\/(.+)$/);
   assert(m);
   return {bucketName: m[1], filename: m[2]};
 }
