+++ mirror/mirror-server/src/functions/app/publish.function.ts
@@ -11,7 +11,7 @@ import * as schema from 'mirror-schema/src/deployment.js';
 import * as semver from 'semver';
 import {newDeploymentID} from 'shared/src/mirror/ids.js';
 import {isSupportedSemverRange} from 'shared/src/mirror/is-supported-semver-range.js';
-import {storeModule} from 'shared/src/mirror/store-module.js';
+import {storeModule} from 'mirror-schema/src/module.js';
 import type {CfModule} from '../../cloudflare/create-worker-upload-form.js';
 import {getServerModuleMetadata} from '../../cloudflare/get-server-modules.js';
 import {publish as publishToCloudflare} from '../../cloudflare/publish.js';
