+++ packages/shared/package.json
@@ -13,7 +13,6 @@
   },
   "devDependencies": {
     "@badrap/valita": "^0.2.0",
-    "@google-cloud/storage": "^6.12.0",
     "@rocicorp/eslint-config": "^0.4.2",
     "@rocicorp/logger": "^5.2.0",
     "@rocicorp/prettier-config": "^0.1.1",
